fix: refactored add transaction modal
