+++ src/components/EditTransaction/EditTransaction.jsx
@@ -4,11 +4,14 @@ import { closeModal } from "../../redux/slices/modal/modalSlice";
 import FormEdit from "../FormEdit/FormEdit";
 import FormExpenseValidation from "../../utils/validations/FormExpenseValidation";
 import FormIncomeValidation from "../../utils/validations/FormIncomeValidation";
+import transactionsApiSlice from "../../redux/slices/api/transactions/transactionsApiSlice";
 
 const EditTransaction = ({ transaction }) => {
   const { income, _id } = transaction;
   const { modals } = useSelector((state) => state.modal);
   const dispatch = useDispatch();
+  const [updateTransaction] =
+    transactionsApiSlice.useUpdateTransactionMutation();
 
   if (!modals[`${_id}edit`]) return null;
   return (
@@ -19,16 +22,18 @@ const EditTransaction = ({ transaction }) => {
         </div>
 
         <div className={styles.typeOfTransHead}>
-          <p style={{ color: `${income && "#24cca7"}` }}>income</p>
+          <p style={{ color: income && "#24cca7" }}>income</p>
           <p className={styles.slashLabel}>/</p>
-          <p style={{ color: `${!income && "#ff6596"}` }}>expense</p>
+          <p style={{ color: !income && "#ff6596" }}>expense</p>
         </div>
 
         <FormEdit
           validationSchema={
             income ? FormIncomeValidation : FormExpenseValidation
           }
-          transaction={transaction}
+          query={updateTransaction}
+          income={income}
+          _id={_id}
         />
 
         <div className={styles.actBtnsWrap}>
