--- src/components/FormExpense/FormExpense.module.scss
@@ -1,63 +1,0 @@
-@import "./../../../public/modern-normalize.scss";
-@import "../../../stylesheet/vars.scss";
-
-.form {
-    margin: 0;
-    padding: 0 11px 0 11px;
-    display: flex;
-    flex-direction: column;
-}
-
-
-
-.input {
-    margin-left: 40px;
-    width: 190px;
-    height: 24px;
-    border: none;
-    outline: none;
-
-    &__comment {
-        height: 52px;
-    }
-
-    &__option {
-        background-color: yellow;
-        border-radius: 50px;
-        border: 2px solid red;
-    }
-}
-
-.calendarWrap {
-    display: flex;
-    align-items: center;
-    margin-left: 40px;
-}
-
-.separatorShort, .separatorLong {
-    margin-top: 8px;
-    margin-bottom: 40px;
-    width: 280px;
-    height: 0;
-    flex-shrink: 0;
-    border-top: 1px solid #E0E0E0;
-
-}
-
-// .separatorShort {
-//     margin-top: 8px;
-// }
-
-// .separatorLong {
-//     margin-top: 60px;
-// }
-
-.calendarIcon {
-    position: fixed;
-    left: 240px;
-    width: 24px;
-    height: 24px;
-    fill: #4A56E2;
-    cursor: pointer;
-}
-
