+++ src/utils/react-select/reactSelectTempFile.js
@@ -1,0 +1,97 @@
+const categoryOptions = [
+  // { value: '', label: 'Select category' },
+  { value: "Products", label: "Products" },
+  { value: "Main expenses", label: "Main expenses" },
+  { value: "Car", label: "Car" },
+  { value: "Self care", label: "Self care" },
+  { value: "Child care", label: "Child care" },
+  { value: "Household products", label: "Household products" },
+  { value: "Education", label: "Education" },
+  { value: "Leisure", label: "Leisure" },
+  { value: "Other expenses", label: "Other expenses" },
+  { value: "Entertainment", label: "Entertainment" },
+];
+
+const inputCategoryStyles = {
+  container: (provided) => ({
+    ...provided,
+    marginLeft: "40px",
+    padding: "0",
+  }),
+  inputContainer: (provided) => ({
+    ...provided,
+    margin: "0",
+    padding: "0",
+  }),
+  input: (provided) => ({
+    ...provided,
+    margin: "0",
+    padding: "0",
+    color: "#000",
+    fontFamily: "Circe",
+    fontSize: "18px",
+    fontStyle: "normal",
+    fontWeight: "400",
+    lineHeight: "normal,",
+  }),
+  selectInput: (provided) => ({
+    ...provided,
+    // margin: '0',
+    // padding: '0',
+    color: "red",
+    // fontFamily: 'Circe',
+    // fontSize: '18px',
+    // fontStyle: 'normal',
+    // fontWeight: '400',
+    // lineHeight: 'normal,'
+  }),
+  valueContainer: (provided) => ({
+    ...provided,
+    margin: "0",
+    padding: "0",
+  }),
+  control: (provided) => ({
+    ...provided,
+    width: "190px",
+    height: "24px",
+    flexShrink: 0,
+    // color: '#BDBDBD',
+    color: "#000",
+    fontFamily: "Circe",
+    fontSize: "18px",
+    fontStyle: "normal",
+    fontWeight: 400,
+    lineHeight: "normal",
+    backgroundColor: "transparent",
+    border: "none",
+    boxShadow: "none",
+  }),
+  menu: (provided) => ({
+    ...provided,
+    width: "280px",
+    // height: '',
+    flexShrink: 0,
+    borderRadius: "20px",
+    background: "rgba(255, 255, 255, 0.70)",
+    boxShadow: "0px 6px 15px 0px rgba(0, 0, 0, 0.10)",
+    backdropFilter: "blur(25px)",
+  }),
+  option: (provided, state) => ({
+    ...provided,
+    fontFamily: "Circe",
+    fontSize: "18px",
+    fontWeight: 400,
+    color: state.isSelected ? "white" : "#BDBDBD", // Możesz dostosować kolory dla wybranych i niezaznaczonych opcji
+  }),
+  singleValue: (provided) => ({
+    ...provided,
+    fontFamily: "Circe",
+    fontSize: "18px",
+    fontWeight: 400,
+    color: "#000",
+  }),
+  indicatorSeparator: (provided) => ({
+    ...provided,
+    display: "none",
+  }),
+};
